TOBORIAL TUTORIAL + Level Edits + Thickened Mechanic Triggers
I thickened the trigger boxes for all the mechanics in the game because multiple people at the playtest told me that there were cases where tobor would slightly hop over the triggers if the player just got out of a bounce. This change was meant to have the mechanics "catch" the player more if that makes sense. Feel free to not merge that if you don't want.